Add: Automatically install adb
Issue: Fix #9
Update: Patch 1.54.06
Info: With this update, users don't have to worry about fiddling with their PATH anymore, the script installs ADB locally just for itself. Of course, the script won't install ADB for those who already have it in their PATH! It'll check if the command `adb` can be run, and only install ADB locally if not.