Fix: Nox not launching correctly
Update: Please take a look at the `#requirements--installation` section in the `README.md` file, as there's a new way to launch Nox now. If you want more info or need help with it, either open an issue or ask in the Discord server.
Update: Grammar and inconsistent terminal output
Issue: Fix #80